feat: :sparkles: automatically guess icons for overview
